tectonic/xetex-swap.h: rename from XeTeXswap.h
