+++ tectonic/xetex-io.c
@@ -6,9 +6,8 @@
 #include "xetex-core.h"
 #include "teckit-c-Engine.h"
 #include "xetex-XeTeXLayoutInterface.h"
-#include "XeTeXswap.h"
+#include "xetex-swap.h"
 #include "xetex-xetexd.h"
-#include "core-bridge.h"
 
 #include <unicode/ubidi.h>
 #include <unicode/ubrk.h>
